+++ .github/CHANGELOG.md
@@ -6,7 +6,11 @@
 
 * Adds the function `tvd_cutoff_bounds` to calculate bounds in the total variation distance between a Fock-truncated and an ideal GBS distribution. [#210](https://github.com/XanaduAI/thewalrus/pull/210)
 
+* Adds function for calculating threshold detection probabilities for Gaussian states with displacement
+  [#220](https://github.com/XanaduAI/thewalrus/pull/220)
+
 * Adds new functions `total_photon_number_distribution` and `characteristic_function` to study properties of the total photon number distribution of a `k` identical lossy squeezers. [#230](https://github.com/XanaduAI/thewalrus/pull/230/)
+
 
 ### Improvements
 
@@ -26,7 +30,8 @@
 
 This release contains contributions from (in alphabetical order):
 
-Jack Brown, Rachel Chadwick, Nicolas Quesada
+Jack Brown, Jake Bulmer, Rachel Chadwick, Stefano Paesani, Nicolas Quesada
+
 
 ---